requirements.txt
numpy==1.23.2
pandas==1.4.4
sklearn==0.0
tensorflow==2.10.0
pickleshare==0.7.5
Pillow==9.2.0
